hackathon-2/phase-II/Full-Stack-Web-Application/backend/pyproject.toml
[project]
name = "taskapp-backend"
version = "0.1.0"
description = "Backend for TASKAPP - Professional Todo Application"
authors = [{name = "Developer", email = "dev@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "sqlmodel>=0.0.22",
    "pydantic>=2.10.0",
    "pydantic[email]",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.20",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "pytest>=8.3.4",
    "httpx>=0.28.0",
    "better-auth>=0.0.24",
    "pyjwt>=2.10.1",
    "slowapi>=0.1.9",
    "aiosqlite>=0.20.0",
    "pytz>=2024.1"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.0",
    "black>=24.10.0",
    "flake8>=7.1.1",
    "mypy>=1.13.0",
    "ruff>=0.8.0",
    "pre-commit>=4.0.0"
]